+++ app.trading.algoritmico.api/src/AppTradingAlgoritmico.Application/Mappings/MappingProfile.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using AppTradingAlgoritmico.Application.DTOs.Expenses;
 using AppTradingAlgoritmico.Application.DTOs.Users;
 using AppTradingAlgoritmico.Domain.Entities;
 
@@ -10,5 +11,9 @@ public class MappingProfile : Profile
     {
         CreateMap<ApplicationUser, UserDto>()
             .ForMember(dest => dest.Roles, opt => opt.Ignore()); // roles injected separately
+
+        CreateMap<Expense, ExpenseDto>();
+        CreateMap<CreateExpenseDto, Expense>();
+        CreateMap<UpdateExpenseDto, Expense>();
     }
 }